connectors/source-sage/pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
version = "0.1.0"
name = "source-sage"
description = "Source implementation for sage."
authors = [ "Airbyte <contact@airbyte.io>",]
license = "MIT"
readme = "README.md"
documentation = "https://docs.airbyte.com/integrations/sources/sage"
homepage = "https://airbyte.com"
repository = "https://github.com/airbytehq/airbyte"
packages = [ { include = "source_sage" }, {include = "main.py" } ]

[tool.poetry.dependencies]
python = "^3.9,<3.12"
airbyte-cdk = "^0"
xmltodict = "^0.14.2"

[tool.poetry.scripts]
source-sage = "source_sage.run:run"

[tool.poetry.group.dev.dependencies]
requests-mock = "*"
pytest-mock = "*"
pytest-html = "*"
pytest = "*"

[tool.pytest.ini_options]
markers = [
    "integration: mark a test as an integration test."
]
